move to notte-eval, add error outputing, add trajectories
