+++ packages/notte-eval/src/notte_eval/data/webvoyager/webvoyager_single.jsonl
@@ -1,1 +1,1 @@
-{"website_name":"Amazon","id":"webvoyager--Amazon--0","question":"Search an Xbox Wireless controller with green color and rated above 4 stars.","url":"https:\/\/www.amazon.com\/","answer":"Xbox Core Wireless Gaming Controller - Velocity Green; 4.7-star"}
+{"website_name":"Wolfram Alpha","id":"webvoyager--Wolfram Alpha--0","question":"derivative of x^2 when x=5.6","url":"https:\/\/www.wolframalpha.com\/","answer":"11.2"}
