+++ packages/notte-eval/src/notte_eval/webvoyager/test_run.py
@@ -3,20 +3,21 @@ import os
 from typing import Any
 
 import pytest
-from bench_types import (  # pyright: ignore[reportImplicitRelativeImport]
+from loguru import logger
+
+from notte_eval.webvoyager.bench_types import (
     BenchmarkTask,
     RunOutput,
     TaskResult,
 )
-from evaluator import EvaluationResponse, Evaluator  # pyright: ignore[reportImplicitRelativeImport]
-from loguru import logger
-from run import (  # pyright: ignore[reportImplicitRelativeImport]
+from notte_eval.webvoyager.evaluator import EvaluationResponse, Evaluator
+from notte_eval.webvoyager.run import (
     evaluate,
     process_output,
     read_tasks,
     run_task_with_session,
 )
-from webvoyager import WebvoyagerEvaluator  # pyright: ignore[reportImplicitRelativeImport]
+from notte_eval.webvoyager.webvoyager import WebvoyagerEvaluator
 
 webvoyager_tasks = read_tasks(
     "packages/notte-eval/src/notte_eval/data/webvoyager/webvoyager_simple.jsonl", 5
@@ -36,17 +37,17 @@ def evaluator(model: str) -> Evaluator:
 @pytest.mark.asyncio
 @pytest.mark.parametrize("task_tuple", webvoyager_tasks)
 async def test_run(task_tuple: tuple[BenchmarkTask, int], evaluator: Evaluator, model: str):
-    try:
-        task = task_tuple[0]
-        run_num = task_tuple[1]
+    task = task_tuple[0]
+    run_num = task_tuple[1]
+    output_dir = f"raw_output_data/{task.id}/"
+    os.makedirs(output_dir, exist_ok=True)
 
+    try:
         resp: RunOutput = await run_task_with_session(task=task, headless=True, model=model)
         out: TaskResult = await process_output(task=task, out=resp)
         eval: EvaluationResponse = await evaluate(evaluator, out)
         logger.info(f"Eval Result: {eval}")
 
-        output_dir = f"raw_output_data/{task.id}/"
-        os.makedirs(output_dir, exist_ok=True)
         output_dict: dict[str, Any] = {
             "task": task.model_dump(),
             "response": out.convert_to_dict,
@@ -56,6 +57,9 @@ async def test_run(task_tuple: tuple[BenchmarkTask, int], evaluator: Evaluator, 
         out.screenshots.get().save(f"{output_dir}{task.id}--{run_num}.webp")
 
         with open(f"{output_dir}output--{run_num}.json", "w") as f:
-            json.dump(output_dict, f)
+            json.dump(output_dict, f, default=str)
     except Exception as e:
         logger.info(f"An exception occured: {e}")
+
+        with open(f"{output_dir}error--{run_num}.txt", "w") as f:
+            _ = f.write(str(e))
